+++ src/main/java/com/carllosmattos/contacts_book/repository/ContactRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.callosmattos.contacts_book.models.Contact;
 
 @Repository
-public interface ContactRepository extends CrudRepository<Contact, Integer> {
+public interface ContactRepository extends CrudRepository<Contact, Long> {
 
 }
